Added Course resources. Resolved Comments.
