+++ app/src/main/java/in/co/rajkumaar/amritarepo/aums/models/CourseData.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright (c) 2020 RAJKUMAR S
+ */
+
+package in.co.rajkumaar.amritarepo.aums.models;
+
+public class CourseData {
+
+    private String id;
+    private String courseCode;
+    private String courseName;
+
+    public CourseData() {
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getCourseCode() {
+        return courseCode;
+    }
+
+    public void setCourseCode(String courseCode) {
+        this.courseCode = courseCode;
+    }
+
+    public String getCourseName() {
+        return courseName;
+    }
+
+    public void setCourseName(String courseName) {
+        this.courseName = courseName;
+    }
+}
